Set BUILD_SUBDIRS to all
Currently when an application that references sc_multi_uart is released the
app_* directories will be excluded from the zip file. If the top level
Makefile references specific apps running then running make in this directory
will not work as the apps won't be there. Using all instead of a specific list
fixes this.